[docs] icon 수정완료
